requirements.txt
fastapi
requests
Pillow
PyGithub
moviepy
uvicorn
